+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * add trait `RedisOps` to support specialized use cases like ping etc.
     * reexport `redis` crate
     * `RedisPool` implements `ConnectionLike`
+    * `RedisPool::ping` was renamed to `RedisPool::ping_nodes` because there is also a simple ping on `RedisOps`
 
 ## 0.25.0 (BREAKING CHANGE)
     * drop connections only on IO errors